Completely Redid Game Logic
Almost entirely rewrote everything to fit the animation style for
the video background. Landings have been implemented.

Need to change the logic forwhen we're in the air. Jumping up
is going to be entirely different than falling. That way
it'll be easier to fall when we go off a landing.